Add tests for black_white and watched
